+++ packages/api-generator/src/worker.js
@@ -1,0 +1,4 @@
+require('ts-node').register({ transpileOnly: true });
+const run = require('./worker.ts');
+
+module.exports = run.default;
